component file folder ducks restructure
