+++ client/components/product/CandyItem.js
