+++ client/components/product/SingleProduct.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import {connect} from 'react-redux'
-import { fetchSingleProduct, addItem } from '../store'
+import { fetchSingleProduct, addItem } from '../../store'
 import CandyItem from './CandyItem'
 import ReviewAvatar from './ReviewAvatar'
 import { List, Container, Divider, Segment } from 'semantic-ui-react'
