+++ client/components/product/candy-item.spec.js
